cifar/requirements.txt
tensorflow-gpu
tensorflow_datasets
tensorflow_addons